modifying tests and page components
