+++ apps/hubspot/src/components/DisplayMessage.tsx
@@ -1,0 +1,24 @@
+import { Flex, Text } from '@contentful/f36-components';
+import { styles } from '../locations/Page.styles';
+import React from 'react';
+
+type MessageProps = {
+  title: string;
+  message: string;
+};
+function DisplayMessage({ title, message }: MessageProps) {
+  return (
+    <Flex
+      flexDirection="column"
+      alignItems="center"
+      justifyContent="center"
+      className={styles.emptyComponentContainer}>
+      <Text fontSize="fontSizeL" fontWeight="fontWeightDemiBold" marginBottom="spacingXs">
+        {title}
+      </Text>
+      <Text fontColor="gray600">{message}</Text>
+    </Flex>
+  );
+}
+
+export default DisplayMessage;
