+++ apps/hubspot/test/locations/Page.spec.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, waitFor, screen, cleanup, fireEvent } from '@testing-library/react';
+import { render, waitFor, screen, cleanup } from '@testing-library/react';
 import { vi, describe, beforeEach, it, expect, afterEach } from 'vitest';
 import Page from '../../src/locations/Page';
 
@@ -92,7 +92,7 @@ describe('Page Location', () => {
       .mockResolvedValueOnce({
         sys: {
           id: 'entry-1',
-          contentType: { sys: { id: 'ct1', name: 'Fruits' } },
+          contentType: { sys: { id: 'Fruits' } },
           updatedAt: new Date(Date.now() - 3 * 60 * 60 * 1000).toISOString(),
           publishedAt: new Date().toISOString(),
         },
@@ -101,7 +101,7 @@ describe('Page Location', () => {
       .mockResolvedValueOnce({
         sys: {
           id: 'entry-2',
-          contentType: { sys: { id: 'ct2', name: 'Animals' } },
+          contentType: { sys: { id: 'Animals' } },
           updatedAt: new Date(Date.now() - 10 * 60 * 60 * 1000).toISOString(),
           publishedAt: undefined,
         },
@@ -122,26 +122,5 @@ describe('Page Location', () => {
       expect(screen.getByText('2')).toBeTruthy();
       expect(screen.getByText('1')).toBeTruthy();
     });
-  });
-
-  it('calls openEntry when Manage fields is clicked', async () => {
-    mockGetConnectedFields.mockResolvedValue({
-      'entry-1': [{ fieldId: 'title', moduleName: 'mod1', updatedAt: '2024-05-01T10:00:00Z' }],
-    });
-    mockCma.entry.get.mockResolvedValueOnce({
-      sys: {
-        id: 'entry-1',
-        contentType: { sys: { id: 'ct1', name: 'Fruits' } },
-        updatedAt: new Date().toISOString(),
-        publishedAt: new Date().toISOString(),
-      },
-      fields: { title: { 'en-US': 'Banana' } },
-    });
-    mockCma.contentType.get.mockResolvedValue({ displayField: 'title' });
-
-    render(<Page />);
-    const btn = await screen.findByRole('button', { name: /Manage fields/i });
-    fireEvent.click(btn);
-    expect(mockNavigator.openEntry).toHaveBeenCalledWith('entry-1');
   });
 });
